+++ packages/expo-dev-menu/ios/Modules/DevMenuInternalModule.swift
@@ -11,15 +11,12 @@ public class DevMenuInternalModule: Module {
     Name("ExpoDevMenuInternal")
 
     // MARK: JavaScript API
-    Constants {
+    Constant("doesDeviceSupportKeyCommands") {
       #if targetEnvironment(simulator)
-      let doesDeviceSupportKeyCommands = true
+      return true
       #else
-      let doesDeviceSupportKeyCommands = false
+      return false
       #endif
-      return [
-        "doesDeviceSupportKeyCommands": doesDeviceSupportKeyCommands
-      ]
     }
 
     AsyncFunction("reload", DevMenuManager.shared.reload)