+++ packages/expo-print/CHANGELOG.md
@@ -10,6 +10,8 @@
 
 ### 💡 Others
 
+- Change Constants to Constant/Property. ([#38926](https://github.com/expo/expo/pull/38926) by [@jakex7](https://github.com/jakex7))
+
 ## 15.0.4 — 2025-08-26
 
 ### 🐛 Bug fixes